+++ NEWS.md
@@ -5,6 +5,7 @@
 * Changed default portfolio optimization algorithm to `NLOPT_LD_SLSQP`
 (Sequential Least-Squares Quadratic Programming). The optimization is no more robust and faster. 
 * Added ability to pass `opts` list to `simulate_scenario()` function with optimization parameters `nloptr` optimization function and thus changing algorithm or its options. The list of possible options and their default values can be obtain with `nloptr::nloptr.print.options()`. 
+* Added ability to pass a single seed value or vector of seeds `simulate_scenario()` function to reproduce the same results for the same scenario, also for Monte Carlo samples.
 * Changed the reordering behavior in the `plot_scenarios()` which does not reorder scenarios anymore.
 * Changed optimal allocation for the last year of a simulation. 
 It is now overwritten with the optimal allocation from the previous year,